backend/pyproject.toml
[project]
name = "olympus"
version = "0.0.1"
dependencies = [
    "SQLAlchemy==2.0.45",
    "psycopg2-binary==2.9.11",
    "httpx==0.28.1",
    "pydantic==2.12.5",
    "ulid-py==1.1.0",
    "pydantic-settings==2.12.0",
    "httpx==0.28.1",
    "fastapi==0.122.0",
    "uvicorn==0.38.0",
    "websockets==15.0.1",
    "python-multipart==0.0.20",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}